+++ src/1-one/contraction-one/model/number-suffix.js
@@ -1,0 +1,17 @@
+// number suffixes that are not units
+const t = true
+export default {
+  'st': t,
+  'nd': t,
+  'rd': t,
+  'th': t,
+  'am': t,
+  'pm': t,
+  'max': t,
+  '°': t,
+  's': t, // 1990s
+  'e': t, // 18e - french/spanish ordinal
+  'er': t, //french 1er
+  'ère': t, //''
+  'ème': t, //french 2ème
+}